pyproject.toml
[project]
name = "retail-data-harmonizer"
version = "1.0.0"
description = "Retail Retail Data Harmonization Demo"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "Snowflake Inc." }]
requires-python = ">=3.12"
dependencies = [
    "typer>=0.24.0",
    "rich>=14.3.2",
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.22",
    "snowflake-snowpark-python>=1.45.0",
    "snowflake-telemetry-python>=0.5.0",
]

[project.scripts]
demo = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "backend"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.1",
    "httpx>=0.28.1",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "D"]
ignore = ["E501", "SIM102", "SIM117", "D203"]  # SIM102/SIM117 intentional for Streamlit; D203 conflicts with D211

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D100", "D104"]  # Module and package docstrings optional in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: fast, hermetic unit tests",
    "integration: tests requiring Snowflake connection",
    "slow: long-running tests (>500ms)",
]

[tool.coverage.run]
source = ["backend", "cli"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 80
